build(single-spa-angular): move internal functions to `single-spa-angular/internals` package
