+++ .eslintrc
@@ -9,7 +9,7 @@
   "rules": {
     "semi": ["error", "always"],
     "indent": "off",
-    "@typescript-eslint/indent": ["error", 2],
+    "@typescript-eslint/indent": false,
     "@typescript-eslint/explicit-function-return-type": false,
     "@typescript-eslint/explicit-member-accessibility": false,
     "@typescript-eslint/no-explicit-any": false,
