+++ src/internals/package.json
@@ -1,0 +1,9 @@
+{
+  "ngPackage": {
+    "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+    "lib": {
+      "entryFile": "src/index.ts",
+      "flatModuleFile": "single-spa-angular-internals"
+    }
+  }
+}
