+++ src/internals/src/index.ts
@@ -1,0 +1,6 @@
+export { SingleSpaAngularOpts, BootstrappedSingleSpaAngularOpts } from './types';
+export {
+  getContainerEl,
+  chooseDomElementGetter,
+  removeApplicationFromDOMIfIvyEnabled,
+} from './dom';
